requirements.txt
influxdb_client
fastapi
nicegui
matplotlib
numpy
pydantic
uvicorn
python-dotenv
sqlalchemy 
psycopg2-binary
